Avoid `(&&String).to_string()` (#5733)
* Document our selected Clippy flags

* Avoid `(&&String).to_string()`

Calling `to_string()` on `&&String` is about 40% slower than calling
`to_string()` on a `&String` or `String` value. Compiler / stdlib
currently do not cooperate well in this matter, so it's better to avoid
the slower code paths.

The change was fully automated by adding a new lint and running `cargo
xtask check --fix`.

* empty: roll the dice 🎲

GitOrigin-RevId: 5107d5676d29f6252c4e42d1e1b10e6d83e256a6