Avoid eager computation when defining Option/Result fallback value (#5731)
* Improve simple strings unwraps

Unfortunately, `unwrap_or_else(|| "".to_string())` is currently not
flagged by the `manual_unwrap_or_default` Clippy rule.

* Avoid eager computation when defining Option/Result fallback value

The `unwrap_or` on `Option` or `Result` should generally be used with constants.
More complex expressions should be lazy-evaluated with `unwrap_or_else`.

This is mostly a performance improvement, but can also avoid failures if
the fallback is not panic-free code.

This enables the `or_fun_call` Clippy check and updates the code.

* Enable unnecessary_lazy_evaluations Clippy lint

It's a counterpart of `or_fun_call` lint.

* Fix new code from main

GitOrigin-RevId: 99af2d18440d97e4bb30b913276f7e7cb0d6946e