Get rid of warehouse-specific column types in favor of single `StdColumn` (#5669)
* Move `StdColumn` and remove derive from proc macro

This is in preparation for the refactor that will remove all special
warehouse structs and consolidate it all in one struct that will use
`AdapterType` to change behavior.

* Port `BigQueryColumn` to `StdColumn`

This commit ports `BigQueryColumn` to `StdColumn`, and gets rid of
`BigQueryColumn` completely (I deleted the file).

It also lays the basic structure that will be used to port other column
types as well.

Some stuff I did in this commit is temporary to keep the program
compiling and working with the 2 models at the same time. For example,
`StdColumn` implements `BaseColumn` and `BaseColumnProperties`. I want
to get rid of those traits but if I did it now the program would no
longer compile. After everything has been ported over, I will do one
last cleanup commit that will fix all of those things.

* Move `BaseColumn` defaults to `StdColumn`

Lots of tests were failing, so for now I'm moving the defaults to the
new implementation.

I noticed there were 2 separate implementations of `numeric_type()`, so
I collapsed all of them under `StdColumnType`, and now the `StdColumn`
has a `as_static()` which returns the column type.

* Make methods in `StdColumnType` statically typed

I contained all of the jinja-types in the `impl Object` block, and all
the methods in `impl StdColumnType` are fully typed. This will make it
nice to unit test later.

* Use `ArgsIter` instead of `ArgParser`

That's it :)

* Port `SnowflakeColumn` to `StdColumn`

This was much easier than BigQuery now that all the machinery is in
place.

* Port `DatabricksColumn` to `StdColumn`

* Port `RedshiftColumn` to `StdColumn`

* Port `PostgresColumn` to `StdColumn`

* Cleanup `BaseColumn` trait in favor of `StdColumn`

- Removed proc macros
- Swapped `Arc<dyn BaseColumn>` with `StdColumn`

* Move `columns/mod.rs` to `columns.rs`

The entire folder is not needed anymore.

* Fix legitimate test failures

This commit makes minor adjustments to fix legitimate test failures:
- `Value::from_serialize` was not behaving as intended since `StdColumn`
  was getting serialized instead of being passed around as an `Object`.
  I added `Into<Value> for StdColumn` to make it compatible with
  `Value::from`
- `api.Column.create` takes in args instead of kwargs, and that was
  causing test failures because it couldn't find the `numeric_precision`
  and `numeric_scale` args, so it was always `None`

* Update replay tests

These tests were updated to conform with the new output. The main change
was that BigQuery columns now serialize a `MODE` attribute and the parse
adapter now uses the proper BigQuery adapter so the types for that
changed as well.

* Changie entry

* Rename `ColumnAdapterType` to `AdapterSpecificColumnAttr`

* Remove TODO

* Add `mode` to `enumerate()` for BQ and fix debug print

* `iter.finish()?` and fix `unimplemented!()` case

GitOrigin-RevId: 43762b6c2590697418c8b44eeca3d692f11fdb0a