BigQuery materialized views (#5378)
* Remove manual table type check in BQ adapter

I replaced it with the `RelationType::from_bigquery_table_type()` that
is being used everywhere else.

* Open source `relation_config`, consolidate `ClusterConfig`

Moved `relation_config` from `dbt-adapters` to
`dbt-fusion-adapters`.

Consolidated `ClusterConfig` with `BigqueryClusterConfig` since
those were virtually the same thing. I made `BigqueryClusterConfig`
implement a `into_fields() -> Vec<String>` method, and I added an
extension trait to implement `try_from_schema()`.

Replaced some `try_from_schema` with `impl TryFrom<&Schema>` where
possible.

NOTE:
I had to remove the dependency from `sdf-frontend`, and I replaced it
with a `__placeholder__` string.

* Refactor `PartitionConfig`

Make it raise `AdapterError` at adapter layer instead of in
`dbt-schemas`.

Make it possible to get it as ref instead of consuming the object. Will
be used later.

* Make `BigqueryMaterializedViewConfig` be trait

Currently `BigqueryMaterializedViewConfig` is a struct that has only one
representation type. I made it a trait "view" and created a concrete
implementation called `BigqueryMaterializedViewConfigRepr` which can be
converted from the arrow `Schema` as we were already doing.

Then, I changed all the code to use the `dyn trait` instead of the
concrete struct, in preparation for future commits where we'll implement
the same trait but with a `DbtModel` backing repr.

Finally, I created a separate `BigqueryMaterializedViewConfigObject`
struct which wraps a `Box<dyn BigqueryMaterializedViewConfig>` to expose
it as a jinja object.

I also refactored how some things are done, like parsing values out of
the `Schema`, and how defaults are constructed.

* Implement `BigqueryMaterializedViewConfigFromDbtModel`

This commit introduces a new concrete implementation of
`BigQueryMaterializedViewConfig` called
`BigqueryMaterializedViewConfigFromDbtModel`.

It holds an inner `DbtModel` as the repr and visualizes it as a
materialized view.

It implements `TryFrom<DbtModel>` which will assert the invariants. If
those invariants are modified at runtime, the trait implementation will
panic.

It also implements a `Into<DbtModel>` so you can upcast it back to a
`DbtModel` without cloning if needed.

* Port `materialized_view.sql` and add Jinja methods

This ports the `materialized_view.sql` from the Core implementation, and
adds the necessary Jinja methods it requires to run.

* Add unit tests to `try_from(Schema)`

I went through each field in the Go ADBC driver and cross checked the
format with what we were doing here.

I fixed two typos, and I also realized we were trying to parse
`RefreshInterval` as `f64` when in reality the driver formats it as
`00h00m00s`, so we're using `parse_duration` from `dbt-xdbc` for that.

* Add changeset for BigQuery materialized views

Implement the changeset that gets passed to the macros.

* Disable `cluster_by` for now

Since we can't pull the status of `cluster_by` on the remote, this logic
was always considering the local status to be different than remote,
which would cause a DROP/CREATE every time you build, even if the remote
is properly clustered by.

For now, I'm disabling it from the changeset until I fix the driver.

* Add record/replay tests

* Changie

* Fix tests

* Make `refresh_interval_minutes` be `Option<f64>`

The docs say its float, but we were parsing it as `u64`. This made
values like `60.0` be parsed as `None` (as seen in previous tests).

I changed it to `f64` now.

* Incorporate PR review feedback

* Skip test

* Fix clippy

* Move consts to top of file

GitOrigin-RevId: 9e545b0c55a631f5a64fa4258ec9823d51dac0ee