Align freshness and event_time merging behavior with dbt-mantle (#5909)
* Add an emptyness check on FreshnessRules

This will assist in truthiness checks against partially empty freshness
configurations.

* Fix freshness logic for overriding top-level configs with table-level configs

* Ensure table-level configs override source-level configs for time_event

* Static analysis and formatting fixes

* Add unit tests for merging event_time and fresheness

* Update changelog

* Simplify pattern matching for table overrides

* Fix regression introduced by https://github.com/dbt-labs/fs/pull/6108

https://github.com/dbt-labs/fs/pull/6108 introduced a bug in the
freshness merging logic that was causing null table-level freshness to
be set to a default value, which was causing source-level logic to be
overwritten by this default value. When table-level freshness is missing,
it should be completely replaced by source-level freshness if it
exists. The change in https://github.com/dbt-labs/fs/pull/6108 was
breaking this behavior when it set the missing table-level freshness to
default values.

GitOrigin-RevId: 2ef02789b6cc96c6941bf73cc628a9ac89743abb